alternate audio track support
consolidate audio track lists retrieved from demux and playlist into a unified list and expose this list through API / notification.
audio tracks from demux are selectable and working fine as before.
audio tracks from playlist are selectable. if selected audio track is not embedded in TS, it will not be played out, as the alternate audio segments are not retrieved yet.
it will come in next step ...